requirements.txt
anyio==3.7.0
exceptiongroup==1.1.1
fastapi==0.97.0
future==0.18.3
idna==3.4
pydantic==1.10.9
python-whois==0.8.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3